+++ parasolr/query.py
@@ -17,10 +17,12 @@ which will automatically initialize a new :class:`parasolr.django.SolrClient`
 if one is not passed in.
 """
 from collections import OrderedDict
-from typing import Dict, List
+import re
+from typing import Any, Dict, List
 
 from parasolr.solr import SolrClient
 from parasolr.solr.client import QueryResponse
+
 
 class SolrQuerySet:
     """A Solr queryset object that allows for object oriented
@@ -37,14 +39,18 @@ class SolrQuerySet:
     filter_qs = []
     field_list = []
     highlight_field = None
-    facet_field = []
+    facet_field_list = []
     facet_opts = {}
     highlight_opts = {}
     raw_params = {}
 
-
     #: by default, combine search queries with AND
     default_search_operator = 'AND'
+
+    #: any value constant
+    ANY_VALUE = '[* TO *]'
+    #: lookup separator
+    LOOKUP_SEP = '__'
 
     def __init__(self, solr: SolrClient):
         # requires solr client so that this version can be django-agnostic
@@ -108,13 +114,15 @@ class SolrQuerySet:
             for key, val in self.highlight_opts.items():
                 query_opts['hl.%s' % key] = val
 
-        if self.facet_field:
+        if self.facet_field_list:
             query_opts.update({
                 'facet': True,
-                'facet.field': self.facet_field
+                'facet.field': self.facet_field_list
             })
             for key, val in self.facet_opts.items():
-                query_opts['facet.%s' % key] = val
+                # use key as is if it starts with "f."
+                # (field-specific facet options); otherwise prepend "facet."
+                query_opts[key if key.startswith('f.') else 'facet.%s' % key] = val
 
         # include any raw query parameters
         query_opts.update(self.raw_params)
@@ -159,15 +167,63 @@ class SolrQuerySet:
         return self.solr.query(**query_opts).facet_counts['facet_fields']
 
     @staticmethod
-    def _lookup_to_filter(key, value) -> str:
-        """Convert keyword argument key=value pair into a Solr filter.
-        Currently only supports simple case of field:value."""
+    def _lookup_to_filter(key: str, value: Any, tag: str='') -> str:
+        """Convert keyword/value argument, with optional lookups separated by
+        ``__``, including: in and exists. Field names should *NOT* include
+        double-underscores by convention. Accepts an optional tag argument
+        to specify an exclude tag as needed.
 
-        # NOTE: as needed, we can start implementing django-style filters
-        # such as __in=[a, b, c] or __range=(start, end)
-        return '%s:%s' % (key, value)
+            Returns: A propertly formatted Solr query string.
+        """
+        # check for a lookup separator and split
+        lookup = ''
+        # format tag for inclusion if tag
+        if tag:
+            tag = '{!tag=%s}' % tag
+        split_key = key.split(SolrQuerySet.LOOKUP_SEP)
+        if len(split_key) == 1:
+            # simple lookup, return key,value pair
+            return '%s%s:%s' % (tag, key, value)
+        # Implementations of Django-style filters such as __in=[a, b, c]
+        # or __range=(start, end)
 
-    def filter(self, *args, **kwargs) -> 'SolrQuerySet':
+        # NOTE: Assuming there is only one LOOKUP_SEP without error handling
+        key, lookup = split_key
+
+        # __in=[a, b, c] filter
+        if lookup == 'in':
+            # value is a list, join with OR logic for all values in list,
+            # treat '' or None values as flagging an exists query
+            not_exists = False
+            if '' in value or None in value:
+                not_exists = True
+            value = list(filter(lambda x: x not in ['', None], value))
+            # if we have a case where the list was just a falsy value
+            # return as if __exists=False
+            if not value:
+                return '%s-%s:%s' % (tag, key, SolrQuerySet.ANY_VALUE)
+            _filter = '%s:(%s)' % (key, ' OR '.join(value))
+            if not not_exists:
+                return '%s%s' % (tag, _filter)
+            else:
+                # This query handles the fact that query syntax does not
+                # support the simpler positive case. Instead, we do a
+                # negative lookup that negates a positive lookup for
+                # all possible values and double-negates a lookup
+                # for any filtered values (thus producing a positive)
+                # The final output is something like:
+                # -(item_type:[* TO *] OR item_type: book OR periodical)
+                return '%s-(%s:%s OR -%s)' % (tag, key, SolrQuerySet.ANY_VALUE,
+                                              _filter)
+
+        # exists=True/False filter
+        if lookup == 'exists':
+            # Look for all possible values, and either negative or not,
+            # depending on the boolean of value.
+            negate = '' if value else '-'
+            return '%s%s%s:%s' % (tag, negate, key, SolrQuerySet.ANY_VALUE)
+
+    def filter(self, *args, tag: str='', **kwargs) -> 'SolrQuerySet':
         """
         Return a new SolrQuerySet with Solr filter queries added.
         Multiple filters can be combined either in a single
@@ -177,20 +233,30 @@ class SolrQuerySet:
             queryset.filter(item_type='person').filter(birth_year=1900)
             queryset.filter(item_type='person', birth_year=1900)
 
-        To provide a filter that should be used in modified, provide
+        You can also search for pre-defined using lookups (in, exists)::
+
+            queryset.filter(item_type__in=['person', 'book'])
+            queryset.filter(item_type__exists=False)
+
+        Tags may be specified for the filter to be used with facet.field
+        exclusions::
+
+            queryset.filter(item_type='person', tag='person')
+
+        To provide a filter that should be used unmodified, provide
         the exact string of your filter query::
 
             queryset.filter('birth_year:[1800 TO *]')
+
+
 
         """
         qs_copy = self._clone()
 
         # any args are treated as filter queries without modification
         qs_copy.filter_qs.extend(args)
-
         for key, value in kwargs.items():
-            qs_copy.filter_qs.append(self._lookup_to_filter(key, value))
-
+            qs_copy.filter_qs.append(self._lookup_to_filter(key, value, tag=tag))
         return qs_copy
 
     def facet(self, *args: str, **kwargs) -> 'SolrQuerySet':
@@ -212,9 +278,33 @@ class SolrQuerySet:
         qs_copy = self._clone()
 
         # cast args tuple to list for consistency with other iterable fields
-        qs_copy.facet_field = list(args)
+        qs_copy.facet_field_list = list(args)
         # add other kwargs to be prefixed in query_opts
         qs_copy.facet_opts.update(kwargs)
+
+        return qs_copy
+
+    def facet_field(self, field: str, exclude: str='', **kwargs) -> 'SolrQuerySet':
+        """
+        Request faceting for a single field. Returns a new SolrQuerySet
+        with Solr faceting enabled and the field added to
+        the list of facet fields.  Any keyword arguments will be set
+        as field-specific facet  configurations.
+
+        ``ex`` will specify a related filter query tag to exclude when
+        generating counts for the facet.
+
+        """
+        qs_copy = self._clone()
+        # append exclude tag if specified
+        qs_copy.facet_field_list.append('{!ex=%s}%s' % (exclude, field)
+                                         if exclude else field)
+        # prefix any keyword args with the field name
+        # (facet. prefix added in query_opts)
+
+        qs_copy.facet_opts.update({
+            'f.%s.facet.%s' % (field, opt) : value
+            for opt, value in kwargs.items()})
 
         return qs_copy
 
@@ -329,9 +419,8 @@ class SolrQuerySet:
         qs_copy.field_list = list(self.field_list)
         qs_copy.highlight_opts = dict(self.highlight_opts)
         qs_copy.raw_params = dict(self.raw_params)
-        qs_copy.facet_field = list(self.facet_field)
+        qs_copy.facet_field_list = list(self.facet_field_list)
         qs_copy.facet_opts = dict(self.facet_opts)
-
 
         return qs_copy